+++ src/test/java/core/PermutationTest.java
@@ -8,7 +8,7 @@ public class PermutationTest {
 
     @Test
     public void testPermutationSortedInput() {
-        int[] input = {0, 1, 2};
+        long[] input = {0, 1, 2};
         Permutation permutation = new Permutation(input);
         while (permutation.hasNext()) {
             System.out.println(Arrays.toString(permutation.next()));
